Re-organized packages and refactorized getting base dir
