+++ src/main/java/qupath/ext/gdcnn/tasks/AnnotationExportTask.java
@@ -1,9 +1,8 @@
-package qupath.ext.tasks;
+package qupath.ext.gdcnn.tasks;
 
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Paths;
 import java.util.List;
 import java.util.Collection;
 
@@ -14,7 +13,7 @@ import javafx.collections.ObservableList;
 import javafx.concurrent.Task;
 import javax.imageio.ImageIO;
 
-import qupath.ext.utils.Utils;
+import qupath.ext.gdcnn.utils.Utils;
 import qupath.lib.common.GeneralTools;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.images.ImageData;
@@ -53,16 +52,12 @@ public class AnnotationExportTask extends Task<Void> {
     protected Void call() throws Exception {
         try {
             Project<BufferedImage> project = qupath.getProject();
-            String outputBaseDir = QP.PROJECT_BASE_DIR;
+            String outputBaseDir = Utils.getBaseDir(qupath);
             if (project != null) {
                 exportAnnotationsProject(project, outputBaseDir);
             } else {
                 ImageData<BufferedImage> imageData = qupath.getImageData();
                 if (imageData != null) {
-                    outputBaseDir = Paths.get(imageData.getServer().getPath()).toString();
-                    // Take substring from the first slash after file: to the last slash
-                    outputBaseDir = outputBaseDir.substring(outputBaseDir.indexOf("file:") + 5,
-                            outputBaseDir.lastIndexOf("/"));
                     exportAnnotations(imageData, outputBaseDir);
                 } else {
                     logger.error("No image or project is open");
