+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,1 +1,1 @@
-qupath.ext.GDCnnExtension
+qupath.ext.gdcnn.GDCnnExtension
