+++ src/main/resources/qupath/ext/gdcnn/ui/tickIcon.png
